component location refactor
