+++ src/App.js
@@ -1,5 +1,5 @@
 
-import Expenses from './components/Expenses';
+import Expenses from './components/Expenses/Expenses';
 
 function App() {
   const expenses = [
